Improve configuration management and GUI clarity
This commit introduces refinements to the management of app settings, primarily in ConfigurationCtrl and ConfigurationService scripts. Changes include adding loader functions, adjusting apps list handling methods, and introducing additional features for word aliases. It also substitutes the use of 'qmService.hideLoader()' with local 'hideLoader()' function for the sake of clarity and granularity. Some modifications in URL paths and error management are included as well.