Added presentation functionality
This commit introduces functionality for a slide presentation in the dfda-1 app. A new JavaScript file, slides-v1.js, has been added to provide the contents of the slides. The PresentationCtrl, a new Angular controller, handles the logic of the presentation, including progressing through slides and integrating various features such as animations, music and speech synthesis. This functionality is linked into the main dfda-1 app through the addition of an ion-slide-box in the presentation.html file. This enriches the user interface experience by offering dynamic content interaction.