Improve presentation visuals and update data slides
The commit pertains to multiple code updates on the presentation display and data contents. Background image and video elements are added to enhance visual appeal. The positioning of the state image is adjusted and the style for data slides is updated. Additionally, changes also include modifications to the slide content, adding more detailed explanation to each slide.