Less weird robot eyes
The CSS properties for designing the robot's eyes have been modified. The position of the eyes has been changed to a slightly different setting. The eye size was kept the same. However, the color gradient of the eyes has been changed to a darker scheme for better contrast and visual impact.