Add caption functionality to voiceover feature
Introduced a new dynamic caption feature to the voiceover function. A new 'caption-container' has been added to the CSS and HTML, along with a new 'setCaption' function in JavaScript. This feature dynamically displays the spoken text as captions, enhancing accessibility by allowing users to both hear and read the spoken content.