Update Presentation Controller and Music Path
Modified the Presentation Controller's functionality to handle music play/pause actions and updated the music file path in qmStates.js. The new implementation in the Presentation Controller accommodates the music parameter changes in qmStates.js which has been updated from a boolean to the mp3 file path "sound/air-of-another-planet-full.mp3". Now, the application uses this path to handle and play the correct background music.