Add support for human representation and talking feature
The update introduces a human representation by adding a new human container, head, body, and mouth in the UI. It also includes the addition of a new talking feature for the human representation supported by speech synthesis. Moreover, CSS styles and an ignore rule for public video directory have been added. A new public CSS file for closed-captioning is incorporated as well.