Update CSS, mute logs and fix video behavior
Various CSS updates were done to improve the layout and animation of human and robot characters. Log messages relating to the actions of 'speechSynthesis' and visualizer types have been commented out in the qmHelpers.js file to declutter the console. Lastly, in presentationCtrl.js, the video behavior during slides presentation was refined including a mechanism to better handle video errors.