Update .gitignore, refactor slides script and improve slide-change handling
Added a new rule in .gitignore file to ignore specific desktop.ini file. The scripts for the slides-v1.js file were cleaned up, particularly in the grammar of the robot's speech, typographical errors in the filenames, and replacing some redundant code constructs with a reusable function. In the presentationCtrl.js, a function for safely setting the slide index was introduced, protecting against potential index-out-of-bounds errors and encapsulating the action of storing the index in local storage.