Update browser extension popup design and functionality
The background.js file has been updated to modify the size and position of the popup. A tracking method for activity was implemented in trackingPopup.js. The popup can now ask questions that users address using a rating system. This process involves the function that updates questions for the variable name and sets listeners for buttons in the popup. A new HTML file for the tracking popup has also been added, including required scripts and button features for the rating system. Existing files have been updated along with the addition of new ones. The changes in the default options.html file were also committed.