Refactor createFdaiUser function
The parameter name in createFdaiUser function was changed from 'clientUserId' to 'yourSystemUserId'. Additionally, the way how client ID and client secret are passed to the API was modified. They are now passed as headers rather than as part of the request body. This change enhances security by preventing sensitive data from being explicitly included in the request body.