Add express app with multiple controllers and services
Files for express app have been added. The express app includes controllers and services for handling requests like analytics, app settings, authentication, connectors and more. A .openapi-generator folder has also been added, it contains an ignore file, eslint config, and a version file. Most of the heavy lifting is done in Request.js where request parameters are handled and sent to the service. The services return responses which are in turn, handled by the controllers.