Update Next.js app's README.md and add dependencies
In this commit, we extend the Getting Started section in the Next.js app's README.md to include the steps on forking and cloning the repository. We also change the dependencies installation instructions from npm to yarn. Additionally, two packages, 'flowise-embed' and 'flowise-embed-react', are added to the package.json file.